+++ src/collectors/mod.rs
@@ -1,4 +1,5 @@
 pub mod diskstats;
+#[cfg(test)]
 pub mod iostat;
 pub mod raid;
 pub mod smart;